+++ bgw-gui/src/jsMain/kotlin/tools/aqua/bgw/elements/CSSBuilder.kt
@@ -127,6 +127,13 @@ internal fun PropertiesBuilder.fontBuilder(componentViewData: UIComponentData) {
   color = componentViewData.font!!.color.unsafeCast<Color>()
 }
 
+internal fun PropertiesBuilder.simpleFontBuilder(componentViewData: UIComponentData) {
+  fontStyle = componentViewData.font!!.fontStyle.let { it.unsafeCast<FontStyle>() }
+  fontWeight = integer(componentViewData.font!!.fontWeight)
+  fontFamily = cssFont(componentViewData.font!!.family)
+  color = componentViewData.font!!.color.unsafeCast<Color>()
+}
+
 internal fun PropertiesBuilder.placeholderFontBuilder(componentViewData: UIComponentData) {
   color = componentViewData.font!!.color.unsafeCast<Color>()
   opacity = number(0.65)